Fix parser errors
